+++ src/interfaces/plugins/keyboardHandler/simpleKeyboardHandler.interface.ts
@@ -16,4 +16,9 @@ export interface SimpleKeyboardHandlerOptions extends KeyboardHandlerOptions
      * Indication whether is alphanumeric search enabled, meaning that when user types alphanumeric keys (outside of live search), select will try to find option starting with searched text and select it
      */
     alphanumericSearch: boolean;
+
+    /**
+     * Indication whether live search events should be processed
+     */
+    liveSearchEvents: boolean;
 }